+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 221:
 
 * Rename to async_base, stable_async_base
 * role_type is in boost/beast/core/role.hpp (API Change)
+* Cleanup in test::stream internals
 
 Actions Required:
