--- doc/qbk/06_websocket/03_client.qbk
@@ -1,67 +1,0 @@
-[/
-    Copyright (c) 2016-2019 Vinnie Falco (vinnie dot falco at gmail dot com)
-
-    Distributed under the Boost Software License, Version 1.0. (See accompanying
-    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-    Official repository: https://github.com/boostorg/beast
-]
-
-[section Handshaking (Clients)]
-
-A WebSocket session begins when a client sends the HTTP/1.1
-[@https://tools.ietf.org/html/rfc7230#section-6.7 Upgrade]
-request for
-[@https://tools.ietf.org/html/rfc6455#section-1.3 websocket],
-and the server sends an appropriate response indicating that
-the request was accepted and that the connection has been upgraded.
-The Upgrade request must include the
-[@https://tools.ietf.org/html/rfc7230#section-5.4 Host]
-field, and the
-[@https://tools.ietf.org/html/rfc7230#section-5.3 target]
-of the resource to request.
-A typical HTTP Upgrade request created and sent by the implementation
-will look like this:
-
-[table WebSocket HTTP Upgrade Request
-[[Serialized Octets][Description]]
-[[
-```
-    GET / HTTP/1.1
-    Host: www.example.com
-    Upgrade: websocket
-    Connection: upgrade
-    Sec-WebSocket-Key: 2pGeTR0DsE4dfZs2pH+8MA==
-    Sec-WebSocket-Version: 13
-    User-Agent: Boost.Beast/216
-```
-][
-    The host and target parameters become part of the Host field
-    and request-target in the resulting HTTP request. The key is
-    generated by the implementation. Callers may add fields or
-    modify fields by providing a ['decorator], described later.
-]]]
-
-The
-[link beast.ref.boost__beast__websocket__stream `stream`]
-member functions
-[link beast.ref.boost__beast__websocket__stream.handshake `handshake`] and
-[link beast.ref.boost__beast__websocket__stream.async_handshake `async_handshake`]
-are used to send the request with the required host and target strings. The
-code below sends the WebSocket HTTP Upgrade request, then reads and processes
-the response:
-
-[code_websocket_3_client_1]
-
-When a client receives an HTTP Upgrade response from the server indicating
-a successful upgrade, the caller may wish to perform additional validation
-on the received HTTP response message. For example, to check that the
-response to a basic authentication challenge is valid. To achieve this,
-overloads of the handshake member function allow the caller to store the
-received HTTP message in an output reference argument of type
-[link beast.ref.boost__beast__websocket__response_type `response_type`]
-as follows:
-
-[code_websocket_3_client_2]
-
-[endsect]
