+++ doc/qbk/06_websocket/06_timeouts.qbk
@@ -1,0 +1,14 @@
+[/
+    Copyright (c) 2016-2019 Vinnie Falco (vinnie dot falco at gmail dot com)
+
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+    Official repository: https://github.com/boostorg/beast
+]
+
+[section Timeouts __new__]
+
+[/-----------------------------------------------------------------------------]
+
+[endsect]
