+++ doc/qbk/main.qbk
@@ -144,8 +144,16 @@
 [import ../../test/doc/core_3_timeouts.cpp]
 [import ../../test/doc/core_4_layers.cpp]
 [import ../../test/doc/http_10_custom_parser.cpp]
+[import ../../test/doc/websocket_common.ipp]
 [import ../../test/doc/websocket.cpp]
-[import ../../test/doc/websocket_3_handshake.cpp]
+[import ../../test/doc/websocket_1_connecting.cpp]
+[import ../../test/doc/websocket_2_handshaking.cpp]
+[import ../../test/doc/websocket_3_decorator.cpp]
+[import ../../test/doc/websocket_4_messages.cpp]
+[import ../../test/doc/websocket_5_control_frames.cpp]
+[import ../../test/doc/websocket_6_timeouts.cpp]
+[import ../../test/doc/websocket_7_teardown.cpp]
+[import ../../test/doc/websocket_8_notes.cpp]
 
 [import ../../include/boost/beast/core/detect_ssl.hpp]
 [import ../../test/beast/core/rate_policy.cpp]
