+++ doc/xsl/class_detail.xsl
@@ -1,6 +1,6 @@
 <!-- CLASS_DETAIL_TEMPLATE BEGIN -->
     <xsl:when test="type = 'class AsyncStream'">
-      <xsl:text>class ``[link beast.concepts.streams.AsyncStream [*AsyncStream]]``</xsl:text>
+      <xsl:text>class __AsyncStream__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class AsyncReadStream'">
       <xsl:text>class __AsyncReadStream__</xsl:text>
@@ -12,11 +12,11 @@
       <xsl:text>class __Body__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class BufferSequence'">
-      <xsl:text>class ``[link beast.concepts.BufferSequence [*BufferSequence]]``</xsl:text>
+      <xsl:text>class __BufferSequence__</xsl:text>
     </xsl:when>
     <xsl:when test="(type = 'class' or type = 'class...') and declname = 'BufferSequence'">
       <xsl:value-of select="type"/>
-      <xsl:text> ``[link beast.concepts.BufferSequence [*BufferSequence]]``</xsl:text>
+      <xsl:text> __BufferSequence__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'CompletionCondition' or type = 'class CompletionCondition'">
       <xsl:text>class __CompletionCondition__</xsl:text>
@@ -34,7 +34,7 @@
       <xsl:text>class __ConstBufferSequence__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'DynamicBuffer' or type = 'class DynamicBuffer'">
-      <xsl:text>class ``[link beast.concepts.DynamicBuffer [*DynamicBuffer]]``</xsl:text>
+      <xsl:text>class __DynamicBuffer__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class EndpointSequence'">
       <xsl:text>class __EndpointSequence__</xsl:text>
@@ -46,7 +46,7 @@
       <xsl:text>class __Executor__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class Fields' or substring(type, 1, 13) = 'class Fields '">
-      <xsl:text>class ``[link beast.concepts.Fields [*Fields]]``</xsl:text>
+      <xsl:text>class __Fields__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'Handler' or type = 'class Handler'">
       <xsl:text>class __Handler__</xsl:text>
@@ -64,16 +64,16 @@
       <xsl:text>class __RangeConnectHandler__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class RatePolicy'">
-      <xsl:text>class ``[link beast.concepts.RatePolicy [*RatePolicy]]``</xsl:text>
+      <xsl:text>class __RatePolicy__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'ReadHandler' or type = 'class ReadHandler'">
       <xsl:text>class __ReadHandler__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'Stream' or type = 'class Stream'">
-      <xsl:text>class ``[link beast.concepts.streams.Stream [*Stream]]``</xsl:text>
+      <xsl:text>class __Stream__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class SyncStream'">
-      <xsl:text>class ``[link beast.concepts.streams.SyncStream [*SyncStream]]``</xsl:text>
+      <xsl:text>class __SyncStream__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'SyncReadStream' or type = 'class SyncReadStream'">
       <xsl:text>class __SyncReadStream__</xsl:text>
