+++ test/doc/websocket_common.ipp
@@ -1,0 +1,20 @@
+//
+// Copyright (c) 2016-2019 Vinnie Falco (vinnie dot falco at gmail dot com)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+// Official repository: https://github.com/boostorg/beast
+//
+
+//[code_websocket_1b
+
+namespace net = boost::asio;
+using namespace boost::beast;
+using namespace boost::beast::websocket;
+
+net::io_context ioc;
+tcp_stream sock(ioc);
+net::ssl::context ctx(net::ssl::context::sslv23);
+
+//]
